Split state/git.rs into scan + fingerprint submodules
