+++ src-tauri/tests/file_mutation_routes.rs
@@ -43,7 +43,7 @@ fn harness() -> Harness {
         version: Arc::new("test".into()),
         event_bus: condash_lib::events::EventBus::default(),
         pty_registry: condash_lib::pty::PtyRegistry::new(),
-        runner_registry: condash_lib::runners::RunnerRegistry::new(),
+        runner_registry: condash_lib::runner_registry::RunnerRegistry::new(),
     };
 
     Harness {
